refactor code to show fav option in movie detail page
