+++ client/src/components/movieCard/index.jsx
@@ -1,84 +1,43 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
-import FontAwesome from "react-fontawesome";
 
 // Components
 import Modal from "../movieModal";
 import MovieDetails from "../movieDetails";
 
 // Styled Components
-import { FavoriteMarker, StyledMovieCard } from "./styles";
-import { connect } from "react-redux";
-import axios from "axios";
-import Actions from "../../store/actions";
+import { StyledMovieCard } from "./styles";
+import FavoriteIcon from "./favoriteIcon";
 
 const MovieCard = (props) => {
-  const { image, movie, clickable, user, favMovie } = props
+  const { image, movie, clickable, favMovie, movieId } = props
   const [isModal, setModal] = useState(false);
     
-    const toggleFavoriteMovie = async (movieId) => {
-        const { updateUser } = props;
-        if (favMovie) {
-            const removeFavRequest = {
-                query: `
-            mutation {
-                removeMovieFromFavorities(email: "${user.profile.email}", movieID: ${movieId}){ 
-                    fav_movies            
-                }
-            }
-        `
-            };
-            const response = await axios.post(process.env.REACT_APP_GRAPHQL_ENDPOINT, removeFavRequest,{
-        headers: {
-          Authorization: `Bearer ${user.token}`
-        }
-      });
-            updateUser({profile: {...user.profile, fav_movies: { ...response.data.data.removeMovieFromFavorities.fav_movies}}})
-        }
-        else {
-            const addFavRequest = {
-                query: `
-                    mutation {
-                        addMovieToFavorities(email: "${user.profile.email}", movieID: ${movieId}){                    
-                            fav_movies
-                        }
-                    }
-            `   
-            };
-            const response = await axios.post(process.env.REACT_APP_GRAPHQL_ENDPOINT, addFavRequest,{
-        headers: {
-          Authorization: `Bearer ${user.token}`
-        }
-      });
-            updateUser({profile: {...user.profile, fav_movies: { ...response.data.data.addMovieToFavorities.fav_movies}}})
-        }
-    };
-
   return (
     <StyledMovieCard>
       <>
         <Modal isVisible={isModal} onClose={() => setModal(false)}>
-          <MovieDetails movie={movie} />
+          <MovieDetails movie={movie} favMovie={favMovie}/>
         </Modal>
         {clickable ? (
             <>
-                {user.token &&
-                    <FavoriteMarker className="card-fav" fav={favMovie} onClick = {() => toggleFavoriteMovie(movie.id, favMovie)}>
-                          <FontAwesome className="far fa-bookmark" name="bookmark" size="1x" color="red"/>
-                    </FavoriteMarker>
-                }
+                <FavoriteIcon movieId={movie.id} favMovie={favMovie} />
                 <img
                     className="clickable"
                     src={image}
                     alt="moviecard"
                     onClick={() => setModal(true)}
+                    
                 />
-                <div className="card-rating">
+                <div className="card-rating" >
                     <div>{movie.vote_average}</div>
                 </div>
           </>
-        ) : (
-          <img src={image} alt="moviecard" />
+              ) : (
+                    <>
+                        <FavoriteIcon movieId={movieId} favMovie={favMovie} />
+                        <img src={image} alt="moviecard" />
+                    </>
         )}
       </>
     </StyledMovieCard>
@@ -91,17 +50,4 @@ MovieCard.propTypes = {
   clickable: PropTypes.bool,
 };
 
-const mapStateToProps = (state) => ({
-  user: state.user
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  updateUser: (...params) => {
-    dispatch(Actions.storeUserProfile.updateUser(...params));
-  },
-});
-
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(MovieCard)
+export default MovieCard;
